% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_bayerisch}
\alias{get_bayerisch}
\title{get_bayerisch}
\usage{
get_bayerisch(txt)
}
\arguments{
\item{txt}{provide a text you want to translate from German into Bavarian (a maximum of 1009 characters per call)}
}
\description{
Translates German to Bavarian
}
